pyproject.toml
[project]
name = "decoder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bulletchess>=0.4.0",
    "numpy>=2.4.0",
    "pyarrow>=22.0.0",
    "python-chess>=1.999",
    "pyyaml>=6.0.3",
    "torch>=2.9.1",
    "torchao>=0.15.0",
    "torchtune>=0.6.1",
    "tqdm>=4.67.1",
    "wandb>=0.23.1",
]
